update add security to check the accessibility role when accessing API

+++ Controllers/CompanyController.cs
@@ -17,10 +17,12 @@ namespace API.Controllers
     public class CompanyController : ControllerBase
     {
         private ICompanyService<Company> _service;
+        private IAuthService _authService;
 
-        public CompanyController(ICompanyService<Company> service)
+        public CompanyController(ICompanyService<Company> service, IAuthService authService)
         {
             _service = service;
+            _authService = authService;
         }
 
         [Authorize]
@@ -29,6 +31,19 @@ namespace API.Controllers
         {
             try
             {
+                var token = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                var user = (User)null;
+                if (token != null)
+                    user = Utils.UserFromToken(token);
+
+                if (user == null)
+                    return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Profil Perusahaan");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
+
                 var total = 0;
                 var result = _service.GetAll(limit, ref page, ref total, search, sort, filter, date);
                 var response = new ListResponse<Company>(result, total, page);
@@ -54,6 +69,19 @@ namespace API.Controllers
         {
             try
             {
+                var token = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                var user = (User)null;
+                if (token != null)
+                    user = Utils.UserFromToken(token);
+
+                if (user == null)
+                    return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Profil Perusahaan");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
+
                 var baseUrl = $"{Request.Scheme}://{Request.Host}";
 
                 var result = _service.GetById(id, baseUrl);
@@ -91,6 +119,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Profil Perusahaan");
+                if (!access.IsCreate)
+                    throw new Exception("Tidak diberikan akses!");
 
                 obj.UserIn = user.UserID.ToString();
                 obj.DateIn = DateTime.Now.AddHours(7);
@@ -132,6 +164,10 @@ namespace API.Controllers
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
 
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Profil Perusahaan");
+                if (!access.IsUpdate)
+                    throw new Exception("Tidak diberikan akses!");
+
                 obj.UserUp = user.UserID.ToString();
 
                 var result = _service.Edit(obj, baseUrl);
@@ -166,6 +202,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Profil Perusahaan");
+                if (!access.IsDelete)
+                    throw new Exception("Tidak diberikan akses!");
 
                 var result = _service.Delete(id, user.UserID.ToString());
