+++ Controllers/LeaveController.cs
@@ -18,10 +18,12 @@ namespace API.Controllers
     public class LeaveController : ControllerBase
     {
         private ILeaveService _service;
+        private IAuthService _authService;
 
-        public LeaveController(ILeaveService service)
+        public LeaveController(ILeaveService service, IAuthService authService)
         {
             _service = service;
+            _authService = authService;
         }
 
         [Authorize]
@@ -38,6 +40,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Cuti");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
 
                 var total = 0;
                 var result = _service.GetAll(limit, ref page, ref total, search, sort, filter, date, user);
@@ -64,6 +70,19 @@ namespace API.Controllers
         {
             try
             {
+                var token = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                var user = (User)null;
+                if (token != null)
+                    user = Utils.UserFromToken(token);
+
+                if (user == null)
+                    return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Cuti");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
+
                 var result = _service.GetById(id);
                 if (result == null)
                     return BadRequest(new { message = "Invalid ID" });
@@ -99,6 +118,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Cuti");
+                if (!access.IsCreate)
+                    throw new Exception("Tidak diberikan akses!");
 
                 obj.UserIn = user.UserID.ToString();
                 obj.DateIn = DateTime.Now.AddHours(7);
@@ -137,6 +160,10 @@ namespace API.Controllers
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
 
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Cuti");
+                if (!access.IsUpdate)
+                    throw new Exception("Tidak diberikan akses!");
+
                 obj.UserUp = user.UserID.ToString();
 
                 var result = _service.Edit(obj);
@@ -171,6 +198,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Cuti");
+                if (!access.IsDelete)
+                    throw new Exception("Tidak diberikan akses!");
 
                 var result = _service.Delete(id, user.UserID.ToString());
