+++ Controllers/ShiftController.cs
@@ -17,10 +17,12 @@ namespace API.Controllers
     public class ShiftController : ControllerBase
     {
         private IService<Shift> _service;
+        private IAuthService _authService;
 
-        public ShiftController(IService<Shift> service)
+        public ShiftController(IService<Shift> service, IAuthService authService)
         {
             _service = service;
+            _authService = authService;
         }
 
         [Authorize]
@@ -29,6 +31,19 @@ namespace API.Controllers
         {
             try
             {
+                var token = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                var user = (User)null;
+                if (token != null)
+                    user = Utils.UserFromToken(token);
+
+                if (user == null)
+                    return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Shift");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
+
                 var total = 0;
                 var result = _service.GetAll(limit, ref page, ref total, search, sort, filter, date);
                 var response = new ListResponse<Shift>(result, total, page);
@@ -54,6 +69,19 @@ namespace API.Controllers
         {
             try
             {
+                var token = HttpContext.Request.Headers["Authorization"].FirstOrDefault()?.Split(" ").Last();
+
+                var user = (User)null;
+                if (token != null)
+                    user = Utils.UserFromToken(token);
+
+                if (user == null)
+                    return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Shift");
+                if (!access.IsRead)
+                    throw new Exception("Tidak diberikan akses!");
+
                 var result = _service.GetById(id);
                 if (result == null)
                     return BadRequest(new { message = "Invalid ID" });
@@ -89,6 +117,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Shift");
+                if (!access.IsCreate)
+                    throw new Exception("Tidak diberikan akses!");
 
                 obj.UserIn = user.UserID.ToString();
                 obj.DateIn = DateTime.Now.AddHours(7);
@@ -128,6 +160,10 @@ namespace API.Controllers
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
 
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Shift");
+                if (!access.IsUpdate)
+                    throw new Exception("Tidak diberikan akses!");
+
                 obj.UserUp = user.UserID.ToString();
 
                 var result = _service.Edit(obj);
@@ -162,6 +198,10 @@ namespace API.Controllers
 
                 if (user == null)
                     return BadRequest(new { message = "Invalid Token" });
+
+                var access = _authService.CheckRoleAccesibility(user.RoleID, "Perusahaan - Shift");
+                if (!access.IsDelete)
+                    throw new Exception("Tidak diberikan akses!");
 
                 var result = _service.Delete(id, user.UserID.ToString());
